+++ fuse.py
@@ -21,9 +21,15 @@ class Fuse():
 	completionSyntax = None
 	buildViews = BuildViewManager()
 	msgManager = MsgManager()
+	startFuseThread = None
+	startFuseThreadExit = False
+	startFuseEvent = threading.Event()
 
 	def __init__(self):
 		self.interop = Interop(self.recv, self.sendHello, self.tryConnect)
+		self.startFuseThread = threading.Thread(target = self.tryConnectThread)
+		self.startFuseThread.daemon = True
+		self.startFuseThread.start()
 
 	def recv(self, msg):
 		try:
@@ -175,21 +181,38 @@ class Fuse():
 			"EventFilter": ""
 		})
 
-	def tryConnect(self):
-		try:				
-			if getSetting("fuse_enabled") == True and not self.interop.isConnected():
-				try:		
-					if os.name == "nt":
-						CREATE_NO_WINDOW = 0x08000000			
-						subprocess.call(["fuse", "daemon", "-b"], creationflags=CREATE_NO_WINDOW)
-					else:
-						subprocess.call(["fuse", "daemon", "-b"])
-				except:
-					traceback.print_exc()
+	fuseStartedCallback = None
 
-				self.interop.connect()
-		except:
-			traceback.print_exc()
+	def tryConnect(self, callback = None):
+		self.fuseStartedCallback = callback
+		self.startFuseEvent.set()
+
+	def tryConnectThread(self):
+		while not self.startFuseThreadExit:
+			try:				
+				if getSetting("fuse_enabled") == True and not self.interop.isConnected():
+					try:		
+						if os.name == "nt":
+							CREATE_NO_WINDOW = 0x08000000			
+							subprocess.call(["fuse", "daemon", "-b"], creationflags=CREATE_NO_WINDOW)
+						else:
+							subprocess.call(["fuse", "daemon", "-b"])
+					except:
+						traceback.print_exc()
+
+					self.interop.connect()
+					if self.fuseStartedCallback is not None:
+						self.fuseStartedCallback()
+
+				self.startFuseEvent.wait()
+				self.startFuseEvent.clear()
+			except:
+				traceback.print_exc()
+
+	def cleanup(self):
+		self.interop.disconnect()
+		self.startFuseThreadExit = True
+		self.startFuseEvent.set()
 
 def plugin_loaded():
 	global gFuse
@@ -203,7 +226,7 @@ def plugin_loaded():
 
 def plugin_unloaded():
 	global gFuse
-	gFuse.interop.disconnect()
+	gFuse.cleanup()
 	gFuse = None
 
 class FuseEventListener(sublime_plugin.EventListener):
@@ -356,7 +379,7 @@ class FusePreview(sublime_plugin.ApplicationCommand):
 		gFuse.tryConnect()
 
 		for path in paths:			
-			subprocess.Popen(["fuse", "preview", "--target=" + type, path])
+			subprocess.Popen(["fuse", "preview", "--target=" + type, "--name=Sublime Text 3", path])
 	
 	def is_visible(self, type, paths = []):
 		if os.name == "nt" and type == "iOS":
@@ -366,6 +389,8 @@ class FusePreview(sublime_plugin.ApplicationCommand):
 
 	def is_enabled(self, type, paths = []):
 		for path in paths:
+			if path == None:
+				return False
 			fileName, fileExtension = os.path.splitext(path)
 			fileExtensionUpper = fileExtension.upper()
 			if fileExtensionUpper != ".UX" and fileExtensionUpper != ".UNOSLN" and fileExtensionUpper != ".UNOPROJ":